+++ data/types/common/test/duration.go
@@ -1,23 +1,23 @@
 package test
 
 import (
-	dataTypesCommon "github.com/tidepool-org/platform/data/types/common"
+	"github.com/tidepool-org/platform/data/types/common"
 	"github.com/tidepool-org/platform/pointer"
 	"github.com/tidepool-org/platform/test"
 )
 
-func NewDuration() *dataTypesCommon.Duration {
-	datum := dataTypesCommon.NewDuration()
-	datum.Units = pointer.FromString(test.RandomStringFromArray(dataTypesCommon.DurationUnits()))
-	datum.Value = pointer.FromFloat64(test.RandomFloat64FromRange(dataTypesCommon.DurationValueRangeForUnits(datum.Units)))
+func NewDuration() *common.Duration {
+	datum := common.NewDuration()
+	datum.Units = pointer.FromString(test.RandomStringFromArray(common.DurationUnits()))
+	datum.Value = pointer.FromFloat64(test.RandomFloat64FromRange(common.DurationValueRangeForUnits(datum.Units)))
 	return datum
 }
 
-func CloneDuration(datum *dataTypesCommon.Duration) *dataTypesCommon.Duration {
+func CloneDuration(datum *common.Duration) *common.Duration {
 	if datum == nil {
 		return nil
 	}
-	clone := dataTypesCommon.NewDuration()
+	clone := common.NewDuration()
 	clone.Units = pointer.CloneString(datum.Units)
 	clone.Value = pointer.CloneFloat64(datum.Value)
 	return clone